+++ messages/ru-RU/app/modules/options.php
@@ -1,7 +1,7 @@
 <?php
 
 return [
-    'Options' => "Опции",
+    'Options' => "Параметры",
     'ID' => "ИД",
     'Section' => "Секция",
     'Param' => "Параметр",
@@ -11,18 +11,34 @@ return [
     'Label and param' => "Название и параметр",
     'Type' => "Тип параметра",
     'Autoload' => "Автозагрузка",
+    'Protected' => "Защищенный",
     'Created at' => "Создано",
     'Updated at' => "Обновлено",
+    'Protected option' => "Защищенный параметр",
     'Attention! In the module settings, autoloading of application parameters is enabled. The ability to delete parameters with autoloading is limited!' => "Внимание! В настройках модуля включена автозагрузка параметров приложения. Возможность удаления параметров с автозагрузкой ограничена!",
     'Error! You cannot delete parameter `{param}` because it is used in the startup.' => "Ошибка! Вы не можете удалить параметр `{param}` поскольку он используется в автозагрузке.",
-    'OK! Parameter `{param}` successfully deleted.' => "Ок! Параметр `{param}` успешно удалён.",
+    'OK! Parameter `{param}` successfully deleted.' => "Ок! Параметр `{param}` был успешно удалён.",
+    'OK! Parameter `{param}` successfully added.' => "Ок! Параметр `{param}` был успешно добавлен.",
+    'OK! Parameter `{param}` successfully edited.' => "Ок! Параметр `{param}` был успешно отредактирован.",
     'An error occurred while deleting a parameter `{param}`.' => "Произошла ошибка при удалении параметра `{param}`.",
+    'An error occurred while adding a parameter `{param}`.' => "Произошла ошибка при добавлении параметра `{param}`.",
+    'An error occurred while editing a parameter `{param}`.' => "Произошла ошибка при редактировании параметра `{param}`.",
+    'Error! You cannot delete a protected from deletion option `{param}`.' => "Ошибка! Вы не можете удалить защищенный от удаления параметр `{param}`.",
+    'Error! You cannot edit a protected option `{param}`.' => "Ошибка! Вы не можете редактировать защищенный параметр `{param}`.",
 
-    'Create option' => "Добавить опцию",
-    'Update option: {name}' => "Редактирование опции: {name}",
+    'Create option' => "Создание параметра",
+    '&larr; Back to list' => "&larr; Назад к списку",
+    'Add new option' => "Новый параметр",
+    'Update option: {name}' => "Редактирование параметра: {name}",
     'Save' => "Сохранить",
     'Update' => "Обновить",
+    'Edit' => "Редактировать",
     'Delete' => "Удалить",
+
+    'Option details' => "Детали опций",
+    'Close' => "Закрыть",
+
+    'Use the symbol «.» as the delimiter of the group and parameter.' => "Используйте символ «.» в качестве разделителя группы и параметра.",
 
     'All types' => "Все типы",
     'Boolean' => "Булево значение",
@@ -34,8 +50,8 @@ return [
     'NULL' => "NULL",
 
     'All modes' => "Всё режимы",
-    'Autoloading' => "Автозагружаемые",
-    'Not loading' => "Не загружаемые",
+    'Autoloading' => "Автозагружаемый",
+    'Not loading' => "Не загружаемый",
 ];
 
 ?>