+++ views/options/create.php
@@ -3,18 +3,19 @@
 use yii\helpers\Html;
 
 /* @var $this yii\web\View */
-/* @var $model wdmg\options\models\Settings */
+/* @var $model wdmg\users\models\Options */
 
 $this->title = Yii::t('app/modules/options', 'Create option');
 $this->params['breadcrumbs'][] = ['label' => Yii::t('app/modules/options', 'Options'), 'url' => ['index']];
 $this->params['breadcrumbs'][] = $this->title;
 ?>
+<div class="page-header">
+    <h1><?= Html::encode($this->title) ?> <small class="text-muted pull-right">[v.<?= $this->context->module->version ?>]</small></h1>
+</div>
 <div class="options-create">
-
-    <h1><?= Html::encode($this->title) ?></h1>
-
     <?= $this->render('_form', [
         'model' => $model,
+        'optionsTypes' => $optionsTypes,
+        'autoloadModes' => $autoloadModes
     ]) ?>
-
 </div>